% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{all_available_quarters}
\alias{all_available_quarters}
\title{Lists all quarters for which there is data available.}
\usage{
all_available_quarters(cache_json = TRUE)
}
\arguments{
\item{cache_json}{TRUE. Whether the JSON files with the raw data should be cached locally.}
}
\value{
A vector with all the available quarters in the format YYYYMM.
}
\description{
Lists all quarters for which there is data available.
}
