% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_quarters_between}
\alias{all_quarters_between}
\title{Lists all quarters between the starting and the end quarter}
\usage{
all_quarters_between(yyyymm_start = 201803, yyyymm_end = 202106)
}
\arguments{
\item{yyyymm_start}{Start calendar quarter for the time series. Accepted formats are: a six-digit integer representing YYYYMM, or a 'Date' class string. Use `NULL` for all available dates. For a list of available series, please use `list_dates`.}

\item{yyyymm_end}{End calendar quarter for the time series. Accepted formats are: a six-digit integer representing YYYYMM, or a 'Date' class string. Use `NULL` for all available dates. For a list of available series, please use `list_dates`.}
}
\description{
Lists all quarters between the starting and the end quarter
}
