% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_backend.R
\name{bank_register}
\alias{bank_register}
\title{Auxiliary function to download registered information about banks and other financial institutions.}
\usage{
bank_register(yyyymm)
}
\arguments{
\item{yyyymm}{WIP adapt inheritParam}
}
\value{
A `tibble` with information about the banks and other financial institutions at the following consolidation levels, in order: prudential conglomerates, financial conglomerates, individual institutions, and foreign exchange institutions.
}
\description{
Auxiliary function to download registered information about banks and other financial institutions.
}
\examples{
bank_register(202003)
bank_register(201812)
}
