% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_frontend.R
\name{bank_time_series}
\alias{bank_time_series}
\title{Retrieves bank-level times series.}
\usage{
bank_time_series(
  banks = NULL,
  consolidation_type = c("prudential conglomerate", "financial conglomerate",
    "individual institutions", "foreign exchange institutions"),
  series = NULL,
  start_quarter = NULL,
  end_quarter = NULL,
  consistent_sample = TRUE
)
}
\arguments{
\item{banks}{The banks for which to download data. Use `NULL` for all banks available. For a list of available banks, please use `list_banks`.}

\item{consolidation_type}{The type of institution to be used. Options are: 'prudential conglomerate' (default), 'financial conglomerate', 'individual institutions' and 'foreign exchange institutions'. More than one option is possible if passed in a vector.}

\item{series}{The series to be downloaded for each bank. Use `NULL` for all series available. For a list of available series, please use `list_series`}

\item{start_quarter}{Start quarter for the time series. Accepted formats are: a six-digit integer representing YYYYMM, or a 'Date' class string. Use `NULL` for all available dates. For a list of available series, please use `list_dates`.}

\item{end_quarter}{End quarter for the time series. Accepted formats are: a six-digit integer representing YYYYMM, or a 'Date' class string. Use `NULL` for all available dates. For a list of available series, please use `list_dates`.}

\item{consistent_sample}{Boolean. If set to `TRUE`, it will return only those banks for which data is available for all quarters in the time series.}
}
\value{
A list containing the function call and a `tibble` with the requested time series. Monetary values are denominated in Brazilian Reais (BRL), except for series that have a different currency code in the name (eg, 'USD_ForeignExposure' is denominated in US Dollars).
}
\description{
Retrieves bank-level times series.
}
\examples{
bank_time_series(banks = c('Bradesco', 'Itau'), consolidation_type = 'prudential conglomerate', series = 'TotalAssets', start_quarter = 201803, end_quarter = 202103)
}
