% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brazilianbanks.R
\docType{package}
\name{brazilianbanks}
\alias{brazilianbanks}
\title{brazilianbanks: Data on Brazilian banks and other regulated financial institutions}
\description{
A user-friendly interface to official datasets of bank-level variables
sourced from the Central Bank of Brazil, in particular the 'IF.data' dataset.
See the website for more information, documentation, examples and the source code.
}
\section{Data sources}{

This package uses two data sources: the 'IF.data' system maintained by the Central Bank of Brazil (BCB in the Portuguese acronym) and banks' Pillar 3 regulatory disclosures, which is regulated by the BCB but maintained and populated by each bank.
}

\section{Functions}{

The functions are `available_quarters()` and `get_bank_stats()`.
}

