% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_GDP_data}
\alias{download_GDP_data}
\title{Downloads GDP data from the Brazilian Institute of Geography and Statistics}
\usage{
download_GDP_data(yyyymm_start, yyyymm_end)
}
\arguments{
\item{yyyymm_start}{Start calendar quarter for the time series. Accepted formats are: a six-digit integer representing YYYYMM, or a 'Date' class string. Use `NULL` for all available dates. For a list of available series, please use `list_dates`.}

\item{yyyymm_end}{End calendar quarter for the time series. Accepted formats are: a six-digit integer representing YYYYMM, or a 'Date' class string. Use `NULL` for all available dates. For a list of available series, please use `list_dates`.}
}
\value{
A data.frame with quarters and the corresponding annual nominal GDP amounts in BRL million up to (and including) each quarter.
}
\description{
Downloads GDP data from the Brazilian Institute of Geography and Statistics
}
