% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_backend.R
\name{download_IFdata_data}
\alias{download_IFdata_data}
\title{Download bank-level data from the BCB 'IF.data' (\url{https://www3.bcb.gov.br/ifdata/})}
\usage{
download_IFdata_data(yyyymm)
}
\arguments{
\item{yyyymm}{A date in the format of four-digit year followed by two-digit month. Months must be either 03, 06, 09 or 12.}
}
\value{
A list of databases for the date chosen. The list has four elements, corresponding to each of the four types of institution-level data available in the 'IF.data' system
}
\description{
Download bank-level data from the BCB 'IF.data' (\url{https://www3.bcb.gov.br/ifdata/})
}
\examples{
download_IFdata_data(202012)
download_IFdata_data(202103)
}
