% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifdata_backend.R
\name{download_IFdata_reports}
\alias{download_IFdata_reports}
\title{Downloads information about the IF.Data reports starting from yyyymm_start and the last available one if yyyymm_end is NULL}
\usage{
download_IFdata_reports(yyyymm_start, yyyymm_end = NULL, cache_json)
}
\arguments{
\item{yyyymm_start}{Start calendar quarter for the time series. Accepted formats are: a six-digit integer representing YYYYMM, or a 'Date' class string. Use `NULL` for all available dates. For a list of available series, please use `list_dates`.}

\item{yyyymm_end}{End calendar quarter for the time series. Accepted formats are: a six-digit integer representing YYYYMM, or a 'Date' class string. Use `NULL` for all available dates. For a list of available series, please use `list_dates`.}

\item{cache_json}{TRUE. Whether the JSON files with the raw data should be cached locally.}
}
\value{
List with
}
\description{
Downloads information about the IF.Data reports starting from yyyymm_start and the last available one if yyyymm_end is NULL
}
