% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifdata_backend.R
\name{download_IFdata_reports_info}
\alias{download_IFdata_reports_info}
\title{Downloads the information about all available dates directly from the BCB website or from cached files}
\usage{
download_IFdata_reports_info(cache_json)
}
\arguments{
\item{cache_json}{TRUE. Whether the JSON files with the raw data should be cached locally.}
}
\value{
List with the full result from the BCB IF.data "relatorios" API endpoint.
}
\description{
Downloads the information about all available dates directly from the BCB website or from cached files
}
