% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifdata_backend.R
\name{download_IFdata_values}
\alias{download_IFdata_values}
\title{Downloads and prepares (eg, renaming columns, etc) IF.Data values for a specific quarter}
\usage{
download_IFdata_values(yyyymm, consolidation_type, var_codes, cache_json)
}
\arguments{
\item{yyyymm}{The quarter for which the data will be downloaded.}

\item{consolidation_type}{One of the four consolidation types in IF.Data: 1 - Prudential conglomerates; 2 - Financial conglomerates; 3 - Financial Institutions; 4 - Foreign exchange institutions}

\item{cache_json}{Boolean. Whether or not will downloaded data be cached locally.}
}
\value{
A `tibble` with the IF.Data values for that quarter.
}
\description{
Downloads and prepares (eg, renaming columns, etc) IF.Data values for a specific quarter
}
