% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifdata_backend.R
\name{download_IFdata_variables}
\alias{download_IFdata_variables}
\title{Downloads information on the IF.Data variable names}
\usage{
download_IFdata_variables(yyyymm, cache_json)
}
\arguments{
\item{yyyymm}{The quarter for which the data will be downloaded.}

\item{cache_json}{Boolean. Whether or not will downloaded data be cached locally.}
}
\description{
Downloads information on the IF.Data variable names
}
