% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifdata_backend.R
\name{excess_capital}
\alias{excess_capital}
\title{Adds a column with each firm's D-SIB status (ie, whether they are one of the domestic systemically important banks) and excess capital}
\usage{
excess_capital(dataframe, yyyymm_start, yyyymm_end)
}
\arguments{
\item{dataframe}{A `tibble` containing the bank-level information.}
}
\value{
A `tibble` with the bank-level information plus the added columns.
}
\description{
Adds a column with each firm's D-SIB status (ie, whether they are one of the domestic systemically important banks) and excess capital
}
