% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifdata_backend.R
\name{getColsFolhas}
\alias{getColsFolhas}
\title{Utility that unnests information on nested columns from the original JSON format}
\usage{
getColsFolhas(colArray)
}
\description{
Utility that unnests information on nested columns from the original JSON format
}
