% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifdata_backend.R
\name{growthrate}
\alias{growthrate}
\title{PROBABLY WILL NOT BE INCLUDED: ITS A GOOD FUNCTION BUT VERY EPXENSIVE FOR A LARGE DATASE. Adds columns with quarter-on-quarter growth for numeric columns}
\usage{
growthrate(dataframe)
}
\arguments{
\item{dataframe}{A `tibble` containing the bank-level information.}
}
\value{
A `tibble` with the bank-level information plus the added columns.
}
\description{
PROBABLY WILL NOT BE INCLUDED: ITS A GOOD FUNCTION BUT VERY EPXENSIVE FOR A LARGE DATASE. Adds columns with quarter-on-quarter growth for numeric columns
}
