% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifdata_backend.R
\name{lag_numericvars}
\alias{lag_numericvars}
\title{Adds columns with (one period) lagged information for numeric columns}
\usage{
lag_numericvars(dataframe)
}
\arguments{
\item{dataframe}{A `tibble` containing the bank-level information.}
}
\value{
A `tibble` with the bank-level information plus the added columns.
}
\description{
Adds columns with (one period) lagged information for numeric columns
}
