% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_frontend.R
\name{list_banks}
\alias{list_banks}
\title{Lists all banks and other financial institutions available for the selected quarter and consolidation type(s).}
\usage{
list_banks(
  consolidation_type = c("prudential conglomerate", "financial conglomerate",
    "individual institutions", "foreign exchange institutions"),
  quarter,
  consistent_sample = TRUE
)
}
\arguments{
\item{consolidation_type}{The type of institution to be used. Options are: 'prudential conglomerate' (default), 'financial conglomerate', 'individual institutions' and 'foreign exchange institutions'. More than one option is possible if passed in a vector.}

\item{quarter}{Accepted formats are: a six-digit integer representing YYYYMM, or a 'Date' class string. Multiple values are accepted if passed as a vector. Use `NULL` for all available dates. For a list of available series, please use `list_dates`.}

\item{consistent_sample}{Boolean. If set to `TRUE`, it will return only those banks for which data is available for all quarters in the time series.}
}
\value{
A list of tibbles, where each element of the list corresponds to a consolidation type.
}
\description{
Lists all banks and other financial institutions available for the selected quarter and consolidation type(s).
}
