% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_frontend.R
\name{list_series}
\alias{list_series}
\title{Lists all time series available for the selected consolidation type(s).}
\usage{
list_series(
  consolidation_type = c("prudential conglomerate", "financial conglomerate",
    "individual institutions", "foreign exchange institutions")
)
}
\arguments{
\item{consolidation_type}{The type of institution to be used. Options are: 'prudential conglomerate' (default), 'financial conglomerate', 'individual institutions' and 'foreign exchange institutions'. More than one option is possible if passed in a vector.}
}
\value{
A list of tibbles, where each element of the list corresponds to a consolidation type.
}
\description{
Lists all time series available for the selected consolidation type(s).
}
