% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifdata_backend.R
\name{loans_share_by_geographical_region}
\alias{loans_share_by_geographical_region}
\title{Adds columns with the loan share of each geographical region}
\usage{
loans_share_by_geographical_region(dataframe)
}
\arguments{
\item{dataframe}{A `tibble` containing the bank-level information.}
}
\value{
A `tibble` with the bank-level information plus the added columns.
}
\description{
Adds columns with the loan share of each geographical region
}
