% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifdata_backend.R
\name{loans_share_by_risk_level}
\alias{loans_share_by_risk_level}
\title{Adds columns with the loan share of each risk level}
\usage{
loans_share_by_risk_level(dataframe)
}
\arguments{
\item{dataframe}{A `tibble` containing the bank-level information.}
}
\value{
A `tibble` with the bank-level information plus the added columns.
}
\description{
Adds columns with the loan share of each risk level
}
