% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path_to_json}
\alias{path_to_json}
\title{Retrieves the path to a JSON file in the local system or in the internet.}
\usage{
path_to_json(
  json_url,
  cache_folder_name = "cache_json",
  cached_file_name = NULL,
  cache_json
)
}
\arguments{
\item{json_url}{the URL to the JSON file}

\item{cache_folder_name}{The local folder where the JSON files are (to be) cached.}

\item{cached_file_name}{The full path to the file name where the JSON file is to be saved locally.}

\item{cache_json}{TRUE. Whether the JSON files with the raw data should be cached locally.}
}
\value{
The path to the JSON file saved locally (if `cache_json` is TRUE) or to its URL if FALSE.
}
\description{
Retrieves the path to a JSON file in the local system or in the internet.
}
