% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifdata_backend.R
\name{prepares_var_names}
\alias{prepares_var_names}
\title{Organises the relationship between variable codes and human-readable names}
\usage{
prepares_var_names(
  yyyymm_start,
  yyyymm_end,
  download_IFdata_reportsreports_info,
  cache_json
)
}
\arguments{
\item{yyyymm_start}{Start calendar quarter for the time series. Accepted formats are: a six-digit integer representing YYYYMM, or a 'Date' class string. Use `NULL` for all available dates. For a list of available series, please use `list_dates`.}

\item{yyyymm_end}{End calendar quarter for the time series. Accepted formats are: a six-digit integer representing YYYYMM, or a 'Date' class string. Use `NULL` for all available dates. For a list of available series, please use `list_dates`.}

\item{cache_json}{TRUE. Whether the JSON files with the raw data should be cached locally.}

\item{reports_info}{A data.frame with information from all available IF.data reports for the selected dates}
}
\description{
Organises the relationship between variable codes and human-readable names
}
