% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifdata_backend.R
\name{stable_segmentation}
\alias{stable_segmentation}
\title{Adds a column with the stable proportionality segment}
\usage{
stable_segmentation(dataframe, last_segment = TRUE)
}
\arguments{
\item{dataframe}{A `tibble` containing the bank-level information.}

\item{last_segment}{Boolean. If TRUE, the stable segmentation is taken from the last period. If FALSE, from the first period.}
}
\value{
A `tibble` with the bank-level information plus the added stable proportionality segment column.
}
\description{
Adds a column with the stable proportionality segment
}
