% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifdata_backend.R
\name{subcols}
\alias{subcols}
\title{Retrieves column information from the JSON files provided by the BCB, especially designed to deal with its nested structured}
\usage{
subcols(x)
}
\arguments{
\item{x}{The list of column (or subcolums) information}
}
\value{
A vector with the relevant information for each column
}
\description{
Retrieves column information from the JSON files provided by the BCB, especially designed to deal with its nested structured
}
