% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_backend.R
\name{all_quarters_between}
\alias{all_quarters_between}
\title{Returns all quarters between two quarters}
\usage{
all_quarters_between(start_quarter, end_quarter)
}
\arguments{
\item{start_quarter}{Start quarter for the time series. Accepted formats are: a six-digit integer representing YYYYMM, or a 'Date' class string. Use `NULL` for all available dates. For a list of available series, please use `list_dates`.}

\item{end_quarter}{End quarter for the time series. Accepted formats are: a six-digit integer representing YYYYMM, or a 'Date' class string. Use `NULL` for all available dates. For a list of available series, please use `list_dates`.}
}
\value{
A vector of quarters in the YYYYMM format.
}
\description{
Returns all quarters between two quarters
}
\examples{
all_quarters_between(201803, 202012)
all_quarters_between(as.Date('2018-03-30'), as.Date('2020-12-31'))
}
