% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifdata_backend.R
\name{excess_capital}
\alias{excess_capital}
\title{Adds a column with each firm's D-SIB status (ie, whether they are one of the domestic systemically important banks) and excess capital}
\usage{
excess_capital(dataframe, yyyymm_start, yyyymm_end)
}
\arguments{
\item{dataframe}{A `tibble` containing the bank-level information.}

\item{yyyymm_start}{Start calendar quarter for the time series. Accepted formats are: a six-digit integer representing YYYYMM, or a 'Date' class string. Use `NULL` for all available dates. For a list of available series, please use `list_dates`.}

\item{yyyymm_end}{End calendar quarter for the time series. Accepted formats are: a six-digit integer representing YYYYMM, or a 'Date' class string. Use `NULL` for all available dates. For a list of available series, please use `list_dates`.}
}
\value{
A `tibble` with the bank-level information plus the added columns.
}
\description{
Adds a column with each firm's D-SIB status (ie, whether they are one of the domestic systemically important banks) and excess capital
}
