% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifdata_backend.R
\name{growthrate}
\alias{growthrate}
\title{Adds columns with quarter-on-quarter growth for numeric columns}
\usage{
growthrate(dataframe)
}
\arguments{
\item{dataframe}{A `tibble` containing the bank-level information.}
}
\value{
A `tibble` with the bank-level information plus the added columns.
}
\description{
Adds columns with quarter-on-quarter growth for numeric columns
}
