% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifdata_backend.R
\name{which_idx}
\alias{which_idx}
\title{Returns the index in the list of reports coming from the BCB API}
\usage{
which_idx(yyyymm)
}
\arguments{
\item{yyyymm}{The quarter of interest in the format YYYYMM}
}
\value{
A scalar that indexes the list of reports
}
\description{
Returns the index in the list of reports coming from the BCB API
}
